requirements.txt
torch>=1.10
torchvision
Pillow
matplotlib
